Uploading my MASH-CPP code from last summer
It makes no sense to keep it in the old repository, specially after ongoing discussions about translating MASH into CPP. This code is mostly classes definitions on how I thought the C++ code should be structured for the MBITES section of the project. There's heavy inheritance usage and there's also a classes UML diagram somewhere (I need to find it).