MBITES - bout restructuring, updates to User's Guide
This was the set of changes precipitated by noticing that boutR should
not be a bout, but a post-prandial resting period.  Herein, behavioral
states are limited to BOMS and search = T/F, so mating and sugar
feeding each have search bouts. The search bout is now generic.  The
only behavioral state transition that can be marked in a bout is from X
to ‘D’. Variables were added for bloodfed, gravid, and starved, which
are used in a new function updateState() that adheres to a more
consistent logic. Two models for oogenesis are now explicitly defined.
Timing is now modular; it takes care of the post-prandial resting
period; it queues mating bouts; and it queues estivation (which is no
longer a bout, but a long resting period).  After this redesign, expect
some bug fixes.